refactor(cli): extract `createPropertyTemplateData` helper
